Split simpleapiview metaclass
